src/test/java/mu/mcb/property/evalution/controller/DocumentControllerTest.java
package mu.mcb.property.evalution.controller;

import static org.mockito.Mockito.when;
import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.get;
import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.jsonPath;
import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.status;

import java.util.Arrays;
import java.util.List;

import org.junit.jupiter.api.BeforeEach;
import org.junit.jupiter.api.Test;
import org.mockito.InjectMocks;
import org.mockito.MockitoAnnotations;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.boot.test.autoconfigure.web.servlet.WebMvcTest;
import org.springframework.boot.test.mock.mockito.MockBean;
import org.springframework.http.MediaType;
import org.springframework.test.web.servlet.MockMvc;

import mu.mcb.property.evalution.dto.DocumentTypeDto;
import mu.mcb.property.evalution.service.DocumentService;

@WebMvcTest(DocumentController.class)
public class DocumentControllerTest {

    @Autowired
    private MockMvc mockMvc;

    @MockBean
    private DocumentService documentService;

    @InjectMocks
    private DocumentController documentController;

    @BeforeEach
    void setUp() {
        MockitoAnnotations.openMocks(this);
    }

    @Test
    void testGetDocumentTypes() throws Exception {
        DocumentTypeDto docType1 = new DocumentTypeDto(1L, "Type1");
        DocumentTypeDto docType2 = new DocumentTypeDto(2L, "Type2");
        List<DocumentTypeDto> documentTypes = Arrays.asList(docType1, docType2);

        when(documentService.getDocumentTypes()).thenReturn(documentTypes);

        mockMvc.perform(get("/api/v1/document/documenttypes")
                .contentType(MediaType.APPLICATION_JSON))
                .andExpect(status().isOk())
                .andExpect(jsonPath("$[0].id").value(1L))
                .andExpect(jsonPath("$[0].name").value("Type1"))
                .andExpect(jsonPath("$[1].id").value(2L))
                .andExpect(jsonPath("$[1].name").value("Type2"));
    }
}